services/embedding-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
qdrant-client
pydantic
python-dotenv
